refactor: move env variables interface to common
